Ignore undojoin error when saving after undo
The changeset in https://github.com/fidian/hexmode/pull/1 allows a user
to only press `u` once to revert to the previous version of the file
after saving. However, if the user presses `u` and then saves, an error
is thrown, since undojoin cannot be called after an undo.

To workaround this issue, we catch and silently ignore this error. This
allows saving after an undo, but two presses of `u` are required to
revert to the previous state.